feat: Update routing in App.jsx and RootLayouts
Refactor the routing in `App.jsx` and `RootLayouts` to change the home page route from "/" to "/Shop". Also, update the route path for the "Home" link in `RootLayouts` to "/Home". This ensures that the home page is set to the "Shop" page and the "Home" link correctly navigates to the "Home" page.

Fixes #123